fixed size issue

+++ fib.c
@@ -5,12 +5,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int fibonacci_iterative(int n) {
+unsigned long int fibonacci_iterative(unsigned long int n) {
     if (n <= 1) {
         return n;
     }
-    int a = 0, b = 1, temp;
-    for (int i = 2; i <= n; i++) {
+    unsigned long int a = 0, b = 1, temp;
+    for (unsigned long int i = 2; i <= n; i++) {
         temp = a + b;
         a = b;
         b = temp;
@@ -18,26 +18,26 @@ int fibonacci_iterative(int n) {
     return b;
 }
 
-int fibonacci_recursive(int n) {
+unsigned long int fibonacci_recursive(unsigned long int n) {
     if (n <= 1) {
         return n;
     }
     return fibonacci_recursive(n - 1) + fibonacci_recursive(n - 2);
 }
 
-int main(int argc, char *argv[]) {
+unsigned  long int main(unsigned long int argc, char *argv[]) {
     
-    // Convert the first argument to an integer
-    int first_num = atoi(argv[1]);
+    // Convert the first argument to an unsigned long unsigned long integer
+    unsigned long int first_num = atoi(argv[1]);
 
     FILE *file = fopen(argv[3], "r");
-    int file_num;
-    fscanf(file, "%d", &file_num);
+    unsigned long int file_num;
+    fscanf(file, "%ld", &file_num);
     fclose(file);
 
-    int sum_of_firstNum_fileNum = first_num + file_num - 1 ;
+    unsigned long int sum_of_firstNum_fileNum = first_num + file_num - 1 ;
 
-    int result;
+    unsigned long int result;
     if (argv[2][0] == 'r') {
         result = fibonacci_recursive(sum_of_firstNum_fileNum);
     } else if (argv[2][0] == 'i') {
@@ -46,7 +46,7 @@ int main(int argc, char *argv[]) {
         return 1;
     }
 
-    printf("%d\n", result);
+   printf("%ld\n", result);
 
 
     return 0;
